+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `is_html()` and `html()` to `Mantle\Http_Client\Response` to check if the
   response is HTML and to return an `HTML` instance for making assertions
   against the HTML content.
+- Added `feed()` method to `Mantle\Http_Client\Response` to parse the body
+  as a feed using SimplePie.
 
 ## v1.9.4